// console/src/generated/PortalQuotaWindowDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PortalQuotaScopeDto } from "./PortalQuotaScopeDto";
import type { PortalQuotaWindowKindDto } from "./PortalQuotaWindowKindDto";

export type PortalQuotaWindowDto = { scope: PortalQuotaScopeDto, window_kind: PortalQuotaWindowKindDto, window_start: number | null, reset_at: number | null, started: boolean, cost_used: string, cost_limit: string, };